refactor: update components and utils folder structure
