--- src/common/index.js
@@ -1,7 +1,0 @@
-import * as utils from './utils';
-import PageLoading from './components/PageLoading';
-
-export {
-  PageLoading,
-  utils,
-};
