+++ src/components/index.js
@@ -1,0 +1,2 @@
+export { default as NotFoundPage } from './NotFoundPage';
+export { default as PageLoading } from './PageLoading';
