+++ src/utils/utils.test.js
@@ -5,7 +5,7 @@ import {
   snakeCaseObject,
   convertKeyNames,
   keepKeys,
-} from './utils';
+} from '.';
 
 describe('modifyObjectKeys', () => {
   it('should use the provided modify function to change all keys in and object and its children', () => {
